+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The public API of this library consists of the public functions declared in
 file [H3Core.java](./src/main/java/com/uber/h3core/H3Core.java), and support
 for the Linux x64 and Darwin x64 platforms.
 
+## Unreleased
+### Fixed
+- Fixed memory management in polyfill with multiple holes.
+
 ## [3.0.1] - 2018-04-30
 ### Added
 - Added release settings.